requirements.txt
asgiref==3.7.2
Django==4.2.3
django_debug_toolbar==3.8.1
Pillow==10.0.0
sqlparse==0.4.4
